make_classifier function implemented
This function returns a classifier-subclass starting from a possible Regressor. For example, make_classifier(sklearn.linear_model.ElasticNet) returns a GenericClassifier class that is a subclass of ElasticNet, where the predict method now returns the sign of the result. Also, the score, if not specified by the user, is the accuracy classification score.