Fix some bugs introduced with Session class, now using Bunch.
The custom Session class required all of the parameters to be specified in the config. In a real usecase, of course, this is not necessary true.
Now the use of the Bunch does not depend on the variables specified in the config.
This is taken care of in the analysis phase, where the appropriate defaults are used.
- Also, note that if you do not specify neither "n_splits" in ma_options or "n_splits_regular" in the main config, the default is the default of ModelAssessment (10). Otherwise, if one of them is 0, then no regular experiments are run. If both are specified, then the one in ma_option has the precedence.